+++ src/form/InputFieldRadioCheckKUI/index.tsx
@@ -1,4 +1,4 @@
-import { AtomDiv, AtomInput } from '../../element';
+import { AtomDiv, AtomInput, AtomInputProps } from '../../element';
 import { ChangeEvent, ReactElement, useState } from 'react';
 
 import { IconRadioClose2, IconRadioOpen2 } from '../../icons';
@@ -10,14 +10,15 @@ export type RadioCheckItem = {
   label: string;
   value: string;
 };
-export type InputFieldRadioCheckKUIProps = Pick<LabelKUIProps, 'align' | 'position'> & {
-  initValue?: string;
-  value?: string;
-  getValue: (value: string) => void;
-  items: RadioCheckItem[];
-  icon?: ReactElement;
-  iconChecked?: ReactElement;
-};
+export type InputFieldRadioCheckKUIProps = Pick<LabelKUIProps, 'position'> &
+  AtomInputProps & {
+    initValue?: string;
+    value?: string;
+    getValue?: (value: string) => void;
+    items: RadioCheckItem[];
+    icon?: ReactElement;
+    iconChecked?: ReactElement;
+  };
 
 export const InputFieldRadioCheckKUI: React.FC<InputFieldRadioCheckKUIProps> = ({
   initValue,
@@ -27,11 +28,12 @@ export const InputFieldRadioCheckKUI: React.FC<InputFieldRadioCheckKUIProps> = (
   icon = <IconRadioOpen2 />,
   iconChecked = <IconRadioClose2 />,
   position = 'right',
-  align,
+  onChange,
+  ...rest
 }: InputFieldRadioCheckKUIProps) => {
   const [selectedRadio, setSelectedRadio] = useState<string | null>(initValue || null);
 
-  const flatValues = items.map((elemento) => elemento.value);
+  const flatValues = items.map((element) => element.value);
   const uniqueValues = Array.from(new Set(flatValues));
 
   if (uniqueValues.length !== flatValues.length)
@@ -39,6 +41,7 @@ export const InputFieldRadioCheckKUI: React.FC<InputFieldRadioCheckKUIProps> = (
 
   const handleRadioChange = (event: ChangeEvent<HTMLInputElement>) => {
     if (getValue) getValue(event.target.value);
+    if (onChange) onChange(event);
     value = event.target.value;
     setSelectedRadio(value);
   };
@@ -49,7 +52,6 @@ export const InputFieldRadioCheckKUI: React.FC<InputFieldRadioCheckKUIProps> = (
         <AtomDiv className="KUI-input-field-radio-check-root" key={index}>
           <LabelKUI
             position={position}
-            align={align}
             htmlFor={`${el.value}-${index}`}
             label={el.label}
             className={`'KUI-input-field-radio-check-label KUI-input-field-radio-check-label-${index}'`}
@@ -57,12 +59,12 @@ export const InputFieldRadioCheckKUI: React.FC<InputFieldRadioCheckKUIProps> = (
             <AtomDiv>
               <AtomInput
                 className="KUI-input-field-radio-check-input"
-                id={`${el.value}-${index}`}
                 key={index}
                 type="radio"
                 value={el.value}
                 checked={selectedRadio === el.value}
                 onChange={handleRadioChange}
+                {...rest}
               />
               {selectedRadio === el.value ? icon : iconChecked}
             </AtomDiv>