+++ src/form/base/PasswordInputKUI/types.tsx
@@ -1,7 +1,7 @@
 import { SerializedStyles } from '@emotion/react';
+import { AtomInputComponentStyled } from './styles';
 
-export type PasswordInputKUIProps = {
-  getValue?: (value: string) => void;
+export type PasswordInputKUIProps = Parameters<typeof AtomInputComponentStyled>[0] & {
   css?: SerializedStyles;
   width?: string;
 };