+++ stories/utily/ModalKUI.stories.tsx
@@ -1,0 +1,18 @@
+import { Meta, StoryObj } from '@storybook/react';
+import { ModalKUI } from 'utily';
+import { ButtonKUI } from 'form';
+
+const meta: Meta<typeof ModalKUI> = {
+  title: 'Utils/Modal',
+  component: ModalKUI,
+};
+export default meta;
+
+type Story = StoryObj<typeof ModalKUI>;
+
+export const Default: Story = {
+  args: {
+    open: true,
+    children: <ButtonKUI>Close</ButtonKUI>,
+  },
+};