+++ stories/utily/ProfileCard.stories.tsx
@@ -1,0 +1,18 @@
+import { Meta, StoryObj } from '@storybook/react';
+import { UserProfileCardKUI } from 'utily';
+
+const meta: Meta<typeof UserProfileCardKUI> = {
+  title: 'Utils/UserProfileCard',
+  component: UserProfileCardKUI,
+};
+export default meta;
+
+type Story = StoryObj<typeof UserProfileCardKUI>;
+
+export const Default: Story = {
+  args: {
+    image: 'Default Image',
+    head: 'Default Heading',
+    body: 'Default Body',
+  },
+};